+++ gcip/core/image.py
@@ -1,32 +1,46 @@
+"""This module represents the Gitlab CI [Image](https://docs.gitlab.com/ee/ci/yaml/#image) keyword.
+
+Use `Image` to specify a Docker image to use for the `gcip.core.job.Job`.
+
+```
+job1.set_image(Image("python"))
+job2.set_image(Image("gcr.io/kaniko-project/executor:debug", entrypoint=[""]))
+```
+"""
 from typing import Dict, List, Union, Optional
 
 
 class Image:
     def __init__(self, name: str, *, entrypoint: Optional[List[str]] = None) -> None:
-        """Creates an object which represents an `image` for a job inside a pipeline.
+        """This module represents the Gitlab CI [Image](https://docs.gitlab.com/ee/ci/yaml/#image) keyword.
+
+        Use `Image` to specify a Docker image to use for the `gcip.core.job.Job`.
 
         Args:
-            name (str): URL where to pull image from incl. image tag.
-            entrypoint (Optional[List[str]]): If set, overwrites the containers entrypoint. Defaults to None.
+            name (str): The fully qualified image name. Could include repository and tag as usual.
+            entrypoint (Optional[List[str]]): Overwrites the containers entrypoint. Defaults to None.
         """
         self._name = name
         self._entrypoint = entrypoint
 
     @property
     def name(self) -> str:
-        """Image name"""
+        """Equals the identical Class argument."""
         return self._name
 
     @property
     def entrypoint(self) -> Optional[List[str]]:
-        """Container `entrypoint`"""
+        """Equals the identical Class argument."""
         return self._entrypoint
 
     def render(self) -> Dict[str, Union[str, List[str]]]:
-        """Returns the rendered object to be attached to the job.
+        """Return a representation of this Image object as dictionary with static values.
+
+        The rendered representation is used by the gcip to dump it
+        in YAML format as part of the .gitlab-ci.yml pipeline.
 
         Returns:
-            Dict[str, str]: The dictionary representation of the image
+            Dict[str, Union[str, List[str]]]: A dictionary prepresenting the image object in Gitlab CI.
         """
         rendered: Dict[str, Union[str, List[str]]] = {}
         rendered["name"] = self._name